+++ src/Core/Credentials/ApplicationProfile.php
@@ -14,13 +14,9 @@ declare(strict_types=1);
 namespace Bitrix24\SDK\Core\Credentials;
 
 use Bitrix24\SDK\Core\Exceptions\InvalidArgumentException;
+use Bitrix24\SDK\Core\Exceptions\UnknownScopeCodeException;
 
-/**
- * Class ApplicationProfile
- *
- * @package Bitrix24\SDK\Core\Credentials
- */
-class ApplicationProfile
+readonly class ApplicationProfile
 {
     private const BITRIX24_PHP_SDK_APPLICATION_CLIENT_ID = 'BITRIX24_PHP_SDK_APPLICATION_CLIENT_ID';
 
@@ -30,31 +26,43 @@ class ApplicationProfile
 
     /**
      * ApplicationProfile constructor.
+     * @throws InvalidArgumentException
      */
-    public function __construct(private readonly string $clientId, private readonly string $clientSecret, private readonly Scope $scope)
+    public function __construct(
+        /**
+         * @var non-empty-string $clientId
+         */
+        public string $clientId,
+        /**
+         * @var non-empty-string $clientSecret
+         */
+        public string $clientSecret,
+        public Scope  $scope)
     {
-    }
+        if (trim($clientId) === '') {
+            throw new InvalidArgumentException('clientId cannot be empty');
+        }
 
-    public function getClientId(): string
-    {
-        return $this->clientId;
-    }
-
-    public function getClientSecret(): string
-    {
-        return $this->clientSecret;
-    }
-
-    public function getScope(): Scope
-    {
-        return $this->scope;
+        if (trim($clientSecret) === '') {
+            throw new InvalidArgumentException('clientSecret cannot be empty');
+        }
     }
 
     /**
-     * @throws \Bitrix24\SDK\Core\Exceptions\InvalidArgumentException
+     * Init Application profile from array
+     *
+     * @param array{
+     *      BITRIX24_PHP_SDK_APPLICATION_CLIENT_ID: string,
+     *      BITRIX24_PHP_SDK_APPLICATION_CLIENT_SECRET: string,
+     *      BITRIX24_PHP_SDK_APPLICATION_SCOPE: string
+     *  } $appProfile
+     *
+     * @throws UnknownScopeCodeException
+     * @throws InvalidArgumentException
      */
     public static function initFromArray(array $appProfile): self
     {
+        // check array keys
         if (!array_key_exists(self::BITRIX24_PHP_SDK_APPLICATION_CLIENT_ID, $appProfile)) {
             throw new InvalidArgumentException(sprintf('in array key %s not found', self::BITRIX24_PHP_SDK_APPLICATION_CLIENT_ID));
         }
@@ -67,10 +75,20 @@ class ApplicationProfile
             throw new InvalidArgumentException(sprintf('in array key %s not found', self::BITRIX24_PHP_SDK_APPLICATION_SCOPE));
         }
 
+        // check array on non-empty values
+        $appProfile = array_map('trim', $appProfile);
+        if ($appProfile[self::BITRIX24_PHP_SDK_APPLICATION_CLIENT_ID] === '') {
+            throw new InvalidArgumentException(sprintf('in array key %s cannot be empty', self::BITRIX24_PHP_SDK_APPLICATION_CLIENT_ID));
+        }
+
+        if ($appProfile[self::BITRIX24_PHP_SDK_APPLICATION_CLIENT_SECRET] === '') {
+            throw new InvalidArgumentException(sprintf('in array key %s cannot be empty', self::BITRIX24_PHP_SDK_APPLICATION_CLIENT_SECRET));
+        }
+
         return new self(
             $appProfile[self::BITRIX24_PHP_SDK_APPLICATION_CLIENT_ID],
             $appProfile[self::BITRIX24_PHP_SDK_APPLICATION_CLIENT_SECRET],
-            new Scope(str_replace(' ', '', explode(',', (string) $appProfile[self::BITRIX24_PHP_SDK_APPLICATION_SCOPE]))),
+            Scope::initFromString($appProfile[self::BITRIX24_PHP_SDK_APPLICATION_SCOPE])
         );
     }
 }